decouple gitpod-protocol from public-api
to prevent rebuilding of components which dependent on gitpod-protocol, but not public-api on an API change
