+++ components/dashboard/BUILD.yaml
@@ -20,6 +20,7 @@ packages:
     deps:
       - components/gitpod-protocol:lib
       - components/public-api/typescript:lib
+      - components/public-api/typescript-common:lib
     config:
       commands:
         build:
