+++ components/dashboard/package.json
@@ -9,6 +9,7 @@
     "@connectrpc/connect-web": "1.1.2",
     "@gitpod/gitpod-protocol": "0.1.5",
     "@gitpod/public-api": "0.1.5",
+    "@gitpod/public-api-common": "0.1.5",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-popover": "^1.0.7",
