+++ components/dashboard/src/service/metrics.ts
@@ -4,7 +4,7 @@
  * See License.AGPL.txt in the project root for license information.
  */
 
-import { MetricsReporter } from "@gitpod/public-api/lib/metrics";
+import { MetricsReporter } from "@gitpod/gitpod-protocol/lib/metrics";
 import { getExperimentsClient } from "../experiments/client";
 import { v4 } from "uuid";
 const commit = require("./config.json").commit;
