+++ components/gitpod-protocol/package.json
@@ -59,8 +59,8 @@
         "exit": true
     },
     "dependencies": {
+        "@bufbuild/protobuf": "^1.3.3",
         "@connectrpc/connect": "1.1.2",
-        "@gitpod/public-api": "0.1.5",
         "@types/react": "17.0.32",
         "abort-controller-x": "^0.4.0",
         "ajv": "^6.5.4",
