+++ components/gitpod-protocol/src/index.ts
@@ -18,4 +18,3 @@ export * from "./teams-projects-protocol";
 export * from "./snapshot-url";
 export * from "./webhook-event";
 export * from "./redis";
-export * from "./auth-providers";
