+++ components/public-api/typescript-common/BUILD.yaml
@@ -1,0 +1,17 @@
+packages:
+  - name: lib
+    type: yarn
+    srcs:
+      - src/**
+      - package.json
+      - tsconfig.json
+    deps:
+      - components/gitpod-protocol:lib
+      - components/public-api/typescript:lib
+    config:
+      packaging: library
+      commands:
+        # leeway executes the build and test step in the wrong order, so we need to call a special script that builds before testing
+        test: ["yarn", "test:leeway"]
+      yarnLock: ${coreYarnLockBase}/../yarn.lock
+      tsconfig: tsconfig.json
