+++ components/public-api/typescript-common/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@gitpod/public-api-common",
+  "version": "0.1.5",
+  "license": "AGPL-3.0",
+  "files": [
+    "lib"
+  ],
+  "exports": {
+    "./lib/*": {
+      "types": "./lib/*.d.ts",
+      "import": "./lib/esm/*.js",
+      "require": "./lib/*.js"
+    }
+  },
+  "scripts": {
+    "build": "yarn run build:cjs && yarn run build:esm",
+    "build:cjs": "tsc",
+    "build:esm": "tsc --module es2015 --outDir ./lib/esm",
+    "test": "mocha './**/*.spec.js' --exclude './node_modules/**' --exclude './lib/esm/**' --exit",
+    "test:leeway": "yarn build && yarn test"
+  },
+  "dependencies": {
+    "@gitpod/gitpod-protocol": "0.1.5",
+    "@gitpod/public-api": "0.1.5",
+    "@connectrpc/connect": "1.1.2",
+    "@bufbuild/protobuf": "^1.3.3"
+  },
+  "devDependencies": {
+    "@types/chai-subset": "^1.3.3",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^16.11.6",
+    "chai": "^4.3.4",
+    "mocha": "^10.2.0",
+    "typescript": "~4.4.2"
+  }
+}
