+++ components/public-api/typescript-common/src/auth-providers.ts
