+++ components/public-api/typescript-common/src/public-api-pagination.ts
