+++ components/public-api/typescript-common/tsconfig.json
@@ -1,0 +1,31 @@
+{
+    "compilerOptions": {
+        "rootDir": "src",
+        "experimentalDecorators": true,
+        "outDir": "lib",
+        "declarationDir": "lib",
+        "lib": [
+            "es6",
+            "esnext.asynciterable",
+            "DOM"
+        ],
+        "strict": true,
+        "noEmitOnError": false,
+        "noUnusedLocals": true,
+        "emitDecoratorMetadata": true,
+        "strictPropertyInitialization": false,
+        "downlevelIteration": true,
+        "module": "CommonJS",
+        "moduleResolution": "node",
+        "target": "es6",
+        "jsx": "react",
+        "sourceMap": true,
+        "declaration": true,
+        "declarationMap": true,
+        "skipLibCheck": true,
+        "useUnknownInCatchVariables": false
+    },
+    "include": [
+        "src"
+    ]
+}
